pyproject.toml
[tool.poetry]
name = "cardinal.py"
version = "3.1.0"
description = "A growing bot for managing a Discord server"
authors = ["FallenWarrior2k <mail@fallenwarrior.me>"]
license = "MIT"
packages = [
    { include = "cardinal", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5"
SQLAlchemy = "^1.3"
alembic = "^1.6.2"
dependency-injector = "^4.32.2"
aioitertools = "^0.7"
markdownify = "^0.7.2"

[tool.poetry.dev-dependencies]
black = {version = "^21.5b1", allow-prereleases = true, extras = ["d"]}
flakehell = "^0.8"
flake8-bugbear = "^21.4.3"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^1.3.0"
isort = "^5"
pep8-naming = "^0.11.1"
pytest = "^6"
pytest-asyncio = "^0.15"
pytest-mock = "^3.6"

[tool.flakehell]
format = "default"

max_line_length = 80
max_complexity = 10

exclude = ["migrations"]

# Workaround for flakehell/flakehell#10
extended_default_ignore = []

[tool.flakehell.plugins]
# E501: Covered by B950 (flake8-bugbear), provides automatic Black compat
# E722: Covered by B001 (flake8-bugbear)
pycodestyle = ["+*", "-E501","-E722"]
pyflakes = ["+*"]
mccabe = ["+*"]
# B901: Forbid return in generators (disagree)
flake8-bugbear = ["+*", "-B901"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pep3101 = ["+*"]
pep8-naming = ["+*"]

[tool.isort]
profile = "black"
multi_line_output = 3

skip = ["migrations"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
